requirements.txt
annotated-types==0.7.0
anyio==4.9.0
aws-sdk-signers==0.0.3
aws_sdk_bedrock_runtime==0.0.2
awscrt==0.27.0
boto3==1.38.5
botocore==1.38.5
click==8.1.8
fastapi==0.115.12
h11==0.16.0
idna==3.10
ijson==3.3.0
jmespath==1.0.1
PyAudio==0.2.14
pydantic==2.11.3
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
Rx==3.2.0
s3transfer==0.12.0
six==1.17.0
smithy-aws-core==0.0.1
smithy-aws-event-stream==0.0.1
smithy-core==0.0.2
smithy-http==0.0.1
smithy-json==0.0.1
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
aiosqlite==0.21.0
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36
bcrypt==5.0.0
python-jose[cryptography]==3.4.0
python-multipart==0.0.20
